+++ migrations/20241228083856_create_transaction.js
@@ -7,7 +7,7 @@ exports.up = async function(knex) {
         CREATE TABLE IF NOT EXISTS transactions (
             transaction_id SERIAL PRIMARY KEY,
             account_id INT NOT NULL,
-            amount BIGINT NOT NULL,
+            amount NUMERIC(10, 2) NOT NULL,
             status VARCHAR(20) DEFAULT 'pending',
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             order_id INT NOT NULL
